@@ -15,45 +15,40 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.2.0+5
+version: 1.2.1+6
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  animated_icon_button: ^0.3.0
-  animations: ^1.1.2
-  catcher: ^0.4.0
-  cupertino_icons: ^1.0.1
-  flutter_colorpicker: ^0.3.4
-  flutter_riverpod: ^0.12.1
-  gallery_saver: ^2.0.3
-  get_it: ^5.0.3
-  image_picker: ^0.6.7+21
-  liquid_swipe: ^1.5.0
-  moor: ^3.4.0
-  package_info: ^0.4.3+2
-  path: ^1.7.0
-  path_provider: ^1.6.27
-  permission_handler: ^5.0.1+1
-  screenshot: ^0.2.0
-  share: ^0.6.5+4
-  shared_preferences: ^0.5.12+4
-  sqlite3_flutter_libs: ^0.3.0
-  url_launcher: ^5.7.10
-  waterfall_flow: ^2.0.5
+  animated_icon_button: ^1.0.0
+  animations: ^2.0.0
+  badges: ^2.0.0-nullsafety.1
+  catcher: ^0.6.3
+  flutter_colorpicker: ^0.4.0
+  flutter_riverpod: ^0.14.0+2
+  get_it: ^6.0.0
+  google_sign_in: ^5.0.2
+  image_picker: ^0.7.3
+  liquid_swipe: ^2.1.0
+  moor: ^4.2.1
+  package_info: ^2.0.0
+  path: ^1.8.0
+  path_provider: ^2.0.1
+  screenshot: ^1.0.0-nullsafety.1
+  share: ^2.0.1
+  shared_preferences: ^2.0.5
+  sqlite3_flutter_libs: ^0.4.1
+  url_launcher: ^6.0.2
+  waterfall_flow: ^3.0.1
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  integration_test:
-    sdk: flutter
   build_runner: null
-  flutter_native_splash: ^0.2.8
-  lint: ^1.5.1
-  moor_generator: ^3.4.1
+  flutter_native_splash: ^1.1.5+1
+  lint: ^1.5.3
+  moor_generator: ^4.2.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
