@@ -1,18 +1,17 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+
 import '../utils/theme.dart';
+import 'shared_prefs_provider.dart';
 
-final _sharedPrefsProvider = FutureProvider<SharedPreferences>((_) {
-  return SharedPreferences.getInstance();
-});
-
-final themeProvider = StateNotifierProvider<ThemeProvider>((ref) {
+final themeProvider =
+    StateNotifierProvider<ThemeProvider, AsyncValue<ThemeType>>((ref) {
   return ThemeProvider(ref.read);
 });
 
 class ThemeProvider extends StateNotifier<AsyncValue<ThemeType>> {
   ThemeProvider(Reader read) : super(const AsyncLoading()) {
-    read(_sharedPrefsProvider.future).then((sharedPrefs) {
+    read(sharedPrefsProvider.future).then((sharedPrefs) {
       _sharedPreferences = sharedPrefs;
       _init();
     });
@@ -20,7 +19,7 @@ class ThemeProvider extends StateNotifier<AsyncValue<ThemeType>> {
 
   static const _themeKey = "theme";
 
-  SharedPreferences _sharedPreferences;
+  late SharedPreferences _sharedPreferences;
 
   void _init() {
     state = AsyncData(stringToTheme(_sharedPreferences.getString(_themeKey)));
