@@ -6,7 +6,7 @@ import '../../../widgets/privacy_statement.dart';
 import 'base_info_widget.dart';
 
 class Others extends StatelessWidget {
-  const Others({Key key}) : super(key: key);
+  const Others({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -17,7 +17,7 @@ class Others extends StatelessWidget {
           children: [
             ListTile(
               title: const Text("Change log"),
-              subtitle: const Text("Changes made to the project"),
+              subtitle: const Text("Changes made to the project."),
               leading: const CircleAvatar(child: Icon(Icons.sticky_note_2)),
               onTap: () => _launchURL(
                 "https://github.com/SirusCodes/Heartry/blob/main/CHANGELOG.md",
@@ -25,7 +25,7 @@ class Others extends StatelessWidget {
             ),
             ListTile(
               title: const Text("Privacy statement"),
-              subtitle: const Text("Our commitment towards privacy"),
+              subtitle: const Text("Our commitment towards privacy."),
               leading: const CircleAvatar(child: Icon(Icons.book)),
               onTap: () {
                 Navigator.push<void>(
@@ -40,7 +40,7 @@ class Others extends StatelessWidget {
                 title: const Text("Version"),
                 subtitle: snapshot.connectionState == ConnectionState.done &&
                         snapshot.hasData
-                    ? Text(snapshot.data.version)
+                    ? Text(snapshot.data!.version)
                     : const Text("Can't get version"),
                 leading: const CircleAvatar(child: Icon(Icons.info)),
                 onTap: () {},
