@@ -6,12 +6,13 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../database/database.dart';
 import '../../../init_get_it.dart';
 import '../../../providers/list_grid_provider.dart';
+import '../../../utils/share_helper.dart';
 import '../../../widgets/share_option_list.dart';
 import '../../reader_screen/reader_screen.dart';
 import '../../writing_screen/writing_screen.dart';
 
 class PoemCard extends StatefulWidget {
-  const PoemCard({Key key, @required this.model}) : super(key: key);
+  const PoemCard({Key? key, required this.model}) : super(key: key);
 
   final PoemModel model;
 
@@ -21,7 +22,7 @@ class PoemCard extends StatefulWidget {
 
 class _PoemCardState extends State<PoemCard>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
+  late AnimationController _controller;
 
   @override
   void initState() {
@@ -148,8 +149,15 @@ class _PoemCardState extends State<PoemCard>
         ),
       ),
       builder: (context) => ShareOptionList(
-        title: widget.model.title,
-        poem: widget.model.poem,
+        onShareAsImage: () => ShareHelper.shareAsImage(
+          context,
+          title: widget.model.title,
+          poem: widget.model.poem,
+        ),
+        onShareAsText: () => ShareHelper.shareAsText(
+          title: widget.model.title,
+          poem: widget.model.poem,
+        ),
       ),
     );
   }
@@ -189,7 +197,7 @@ class _PoemCardState extends State<PoemCard>
   Future<void> _delete() async {
     final result = await locator<Database>().deletePoem(widget.model);
 
-    final String msg = result == null ? "Failed to delete" : "Deleted";
+    final String msg = result == 0 ? "Failed to delete" : "Deleted";
 
     Navigator.pop(context);
 
